+++ crates/nostr/CHANGELOG.md
@@ -41,7 +41,7 @@
 - Add NIP-C7 support (https://github.com/rust-nostr/nostr/pull/1067)
 - Add NIP-60 support (https://github.com/rust-nostr/nostr/pull/1092)
 - Implement `ToBech32` trait for `Nip21`
-- Add `CowTag::content`
+- Add `CowTag::kind` and `CowTag::content`
 - Add `Kind::CashuNutZapInfo` (10019) and `Kind::CashuNutZap` (9321) variants
 - Add nip47 holdinvoice methods and notification (https://github.com/rust-nostr/nostr/pull/1019)
 - Add `TransactionState` to `LookupInvoiceResponse` and `PaymentNotification` (https://github.com/rust-nostr/nostr/pull/1045)